removed karhoobookingrequrestpresenter for good
