+++ KarhooUISDK/Screens/BookingRequestScreen/FormBookingRequestScreen/FormBookingRequestPresenter.swift
@@ -18,6 +18,19 @@ final class FormBookingRequestPresenter: BookingRequestPresenter {
     private let tripService: TripService
     private let userService: UserService
     private let bookingMetadata: [String: Any]?
+    private let paymentNonceProvider: PaymentNonceProvider
+    
+    private let analytics: Analytics
+    private let appStateNotifier: AppStateNotifierProtocol
+    private var trip: TripInfo?
+    private var flightNumber: String?
+    private var comments: String?
+    private var bookingRequestInProgress: Bool = false
+    private var flightDetailsScreenIsPresented: Bool = false
+    private let flightDetailsScreenBuilder: FlightDetailsScreenBuilder
+    private let baseFareDialogBuilder: PopupDialogScreenBuilder
+
+    var karhooUser: Bool = false
 
     init(quote: Quote,
          bookingDetails: BookingDetails,
@@ -25,42 +38,187 @@ final class FormBookingRequestPresenter: BookingRequestPresenter {
          threeDSecureProvider: ThreeDSecureProvider = BraintreeThreeDSecureProvider(),
          tripService: TripService = Karhoo.getTripService(),
          userService: UserService = Karhoo.getUserService(),
+         analytics: Analytics = KarhooAnalytics(),
+         appStateNotifier: AppStateNotifierProtocol = AppStateNotifier(),
+         flightDetailsScreenBuilder: FlightDetailsScreenBuilder = KarhooUI().screens().flightDetails(),
+         baseFarePopupDialogBuilder: PopupDialogScreenBuilder = UISDKScreenRouting.default.popUpDialog(),
+         paymentNonceProvider: PaymentNonceProvider = PaymentFactory().nonceProvider(),
          callback: @escaping ScreenResultCallback<TripInfo>) {
         self.threeDSecureProvider = threeDSecureProvider
         self.tripService = tripService
         self.callback = callback
         self.userService = userService
+        self.paymentNonceProvider = paymentNonceProvider
+        self.appStateNotifier = appStateNotifier
+        self.analytics = analytics
+        self.flightDetailsScreenBuilder = flightDetailsScreenBuilder
+        self.baseFareDialogBuilder = baseFarePopupDialogBuilder
         self.quote = quote
         self.bookingDetails = bookingDetails
         self.bookingMetadata = bookingMetadata
+        
+        appStateNotifier.register(listener: self)
     }
 
-    func load(view: BookingRequestView) {
+    func load(view: BookingRequestView, karhooUser: Bool = false) {
         self.view = view
+        self.karhooUser = karhooUser
+        if karhooUser {
+            finishLoad(view: view)
+        }
         view.set(quote: quote)
         setUpBookingButtonState()
         threeDSecureProvider.set(baseViewController: view)
     }
+
+     private func finishLoad(view: BookingRequestView) {
+         if quote.source == .market {
+             view.set(price: CurrencyCodeConverter.quoteRangePrice(quote: quote))
+         } else {
+             view.set(price: CurrencyCodeConverter.toPriceString(quote: quote))
+         }
+
+         view.set(quoteType: quote.quoteType.description)
+         view.set(baseFareExplanationHidden: quote.quoteType == .fixed)
+         paymentNonceProvider.set(baseViewController: view)
+         configureQuoteView()
+         view.paymentView(hidden: userService.getCurrentUser()?.paymentProvider?.provider.type == .adyen)
+     }
     
     func isKarhooUser() -> Bool {
-        false
+        return karhooUser
     }
 
     func bookTripPressed() {
-        view?.setRequestingState()
-        
-        guard let passengerDetails = view?.getPassengerDetails(), let nonce = getPaymentNonceAccordingToAuthState() else {
-            view?.setDefaultState()
-            return
-        }
-
-        if userService.getCurrentUser()?.paymentProvider?.provider.type == .braintree {
-            threeDSecureNonceThenBook(nonce: nonce,
-                                      passengerDetails: passengerDetails)
+        if karhooUser {
+            submitBooking()
         } else {
-            book(threeDSecureNonce: nonce, passenger: passengerDetails)
+            view?.setRequestingState()
+            
+            guard let passengerDetails = view?.getPassengerDetails(), let nonce = getPaymentNonceAccordingToAuthState() else {
+                view?.setDefaultState()
+                return
+            }
+            
+            if userService.getCurrentUser()?.paymentProvider?.provider.type == .braintree {
+                threeDSecureNonceThenBook(nonce: nonce,
+                                          passengerDetails: passengerDetails)
+            } else {
+                book(threeDSecureNonce: nonce, passenger: passengerDetails)
+            }
         }
     }
+
+    private func submitBooking() {
+        bookingRequestInProgress = true
+        view?.setRequestingState()
+        
+        guard let currentUser = userService.getCurrentUser(),
+              let currentOrg = userService.getCurrentUser()?.organisations.first else {
+            view?.showAlert(title: UITexts.Errors.somethingWentWrong,
+                            message: UITexts.Errors.getUserFail,
+                            error: nil)
+            return
+        }
+        
+        if let destination = bookingDetails.destinationLocationDetails {
+            analytics.bookingRequested(destination: destination,
+                                       dateScheduled: bookingDetails.scheduledDate,
+                                       quote: quote)
+        }
+        
+        if let nonce = view?.getPaymentNonce() {
+            bookTrip(withPaymentNonce: Nonce(nonce: nonce), user: currentUser)
+        } else {
+            paymentNonceProvider.getPaymentNonce(user: currentUser,
+                                                 organisation: currentOrg,
+                                                 quote: quote) { [weak self] result in
+                switch result {
+                case .completed(let result): handlePaymentNonceProviderResult(result)
+                case .cancelledByUser: self?.view?.setDefaultState()
+                }
+            }
+        }
+        func handlePaymentNonceProviderResult(_ paymentNonceResult: PaymentNonceProviderResult) {
+             switch paymentNonceResult {
+             case .nonce(let nonce): bookTrip(withPaymentNonce: nonce, user: currentUser)
+             case .cancelledByUser:
+                 self.view?.setDefaultState()
+             case .failedToAddCard(let error):
+                 self.view?.setDefaultState()
+                 self.view?.show(error: error)
+             default:
+                 self.view?.showAlert(title: UITexts.Generic.error, message: UITexts.Errors.somethingWentWrong,
+                                      error: nil)
+                 view?.setDefaultState()
+             }
+        }
+    }
+    
+    private func bookTrip(withPaymentNonce nonce: Nonce, user: UserInfo) {
+        var tripBooking = TripBooking(quoteId: quote.id,
+                                      passengers: Passengers(additionalPassengers: 0,
+                                                             passengerDetails: [PassengerDetails(user: user)]),
+                                      flightNumber: flightNumber)
+        
+        tripBooking.paymentNonce = nonce.nonce
+        
+        var map: [String: Any] = [:]
+        if let metadata = bookingMetadata {
+            map = metadata
+        }
+        tripBooking.meta = map
+        if userService.getCurrentUser()?.paymentProvider?.provider.type == .adyen {
+            tripBooking.meta["trip_id"] = nonce.nonce
+        }
+        
+        tripService.book(tripBooking: tripBooking)
+            .execute(callback: { [weak self] (result: Result<TripInfo>) in
+                self?.bookingRequestInProgress = false
+                self?.handleBookTrip(result: result)
+            })
+    }
+    
+    private func handleBookTrip(result: Result<TripInfo>) {
+         guard let trip = result.successValue() else {
+             view?.setDefaultState()
+
+             if result.errorValue()?.type == .couldNotBookTripPaymentPreAuthFailed {
+                 view?.retryAddPaymentMethod()
+             } else {
+                 callback(ScreenResult.failed(error: result.errorValue()))
+             }
+
+             return
+         }
+
+         self.trip = trip
+         view?.showBookingRequestView(false)
+     }
+    
+    private func configureQuoteView() {
+         if bookingDetails.isScheduled {
+             configurePrebookState()
+             return
+         }
+         configureForAsapState()
+     }
+
+     private func configurePrebookState() {
+         guard let timeZone = bookingDetails.originLocationDetails?.timezone() else {
+             return
+         }
+
+         let prebookFormatter = KarhooDateFormatter(timeZone: timeZone)
+
+         view?.setPrebookState(timeString: prebookFormatter.display(shortStyleTime: bookingDetails.scheduledDate),
+                                 dateString: prebookFormatter.display(mediumStyleDate: bookingDetails.scheduledDate))
+     }
+
+     private func configureForAsapState() {
+         view?.setAsapState(qta: QtaStringFormatter().qtaString(min: quote.vehicle.qta.lowMinutes,
+                                                                  max: quote.vehicle.qta.highMinutes))
+     }
 
     private func getPaymentNonceAccordingToAuthState() -> String? {
         switch Karhoo.configuration.authenticationMethod() {
@@ -147,6 +305,19 @@ final class FormBookingRequestPresenter: BookingRequestPresenter {
     private func setUpBookingButtonState() {
         if TripInfoUtility.isAirportBooking(bookingDetails) {
             view?.setAddFlightDetailsState()
+        }
+    }
+    
+    deinit {
+        appStateNotifier.remove(listener: self)
+    }
+}
+
+extension FormBookingRequestPresenter: AppStateChangeDelegate {
+    func appDidEnterBackground() {
+        if bookingRequestInProgress == false,
+           flightDetailsScreenIsPresented == false {
+            self.didPressClose()
         }
     }
 }
